Update array.py to have correct self.__div__
Fixed __truediv__ functionality.  Submitted this as a patch to the main branch with the following note:

Added basic element-wise addition (+), subtraction (-), multiplication (*), and division (/) to BoltSparkArrays.

Method used is `_rdd().join().mapValues()`.  Attempted to play around with `_rdd().cogroup().reduceByKey()` but `cogroup()` returns interables and I was not familiar enough with them to spend the time troubleshooting other methods for a time efficiency comparison.  May be a better method to implement these operations, but I went with the best I knew how.

I didn't include any of the reverse functions (__radd__ etc) as I wasn't sure how it would behave/interact with local array functions if there was an attempt to list the local array first (`local + Spark`). There are no issues as long as it is always two arrays being added together and the Spark array is listed first (`Spark + local`).

I also didn't add support for adding/subtracting/multiplying/dividing by a constant integer across the whole array.  I wasn't sure if this is functionality we would want to add, as it can be done using `map()`.  My personal opinion is to require the user to use `map()` because I feel it forces the user to remember they are working on distributed arrays.  However, this functionality could be easily implemented into the existing functions with some if statements in checking for integers and doing the `map()` under the hood.

On a similar note to above, these functions could also be changed to not use the symbolic values (+, -, *, /) but be calls required in a map format (`.add()`, `.subtract()`, `multiply()`, `divide()`).

Finally, if there is an intention to add many operations for distributed BoltSparkArrays similar to all the functions for python ndarrays, perhaps an independent class or .py file should be created to contain these operations that could then be called?  I wasn't sure of the long-term intentions of what types of functions there is an interest to integrate into BoltArrays.

Tried to include similar safety catches as `concatenate()`.

For division, true division from __future__ is always performed.

Basic troubleshooting was performed without any major issues, but more thorough troubleshooting may be useful.